Fixes table tags getting stripped
Table output was being suppressed through the  helper which calls  by default. Since  and all its descendants weren't in the whitelist, the tags were getting stripped in the view (and not a Nokogiri issue). This fixes this chain fail by setting the  flag to false and skipping this step since there is no user input.